Add -Wconversion
Summary:
Enable -Wconversion (but not -Wsign-conversion... yet)

Verified that still compiles on OSX too.

Test Plan: make all

Reviewers: rocketspeed, stupaq-fb

Reviewed By: stupaq-fb

Subscribers: stupaq-fb

Differential Revision: https://reviews.facebook.net/D34455

+++ src/client/client.cc
@@ -590,7 +590,8 @@ Statistics ClientImpl::GetStatistics() const {
 }
 
 int ClientImpl::GetWorkerForTopic(const Topic& name) const {
-  return MurmurHash2<std::string>()(name) % msg_loop_->GetNumWorkers();
+  return static_cast<int>(MurmurHash2<std::string>()(name) %
+                          msg_loop_->GetNumWorkers());
 }
 
 MessageReceivedClient::~MessageReceivedClient() {
